requirements.txt
pytorch
numpy
scikit-image
cv2
tqdm
